Generated by Spark: Can we create a "Customer Details Page" when we click on a customer - it will show us thier profile, this will concude their contact information and job history
Can we create a customers page?

can we create a Login page with consistent styling for me to access this from?

From the jobs page i like how when we click a job it opens full screen.

From the quotes page i like how the jobs dropdown inline.

Can we reverse this? ideally in the quotes page they will pop out full screen. ideally in the jobs page the details will dropdown in line.