Generated by Spark: the login page is perfect, can we temporarily hide it so i can see the oher pages?
can we please enable customer grouping by teir?

can we add esport list to CSV

can we add from the "job detail" &  "quote detail" a dropdown button with more actions list. on this list we will be able to navigate to the customers profile, we will be able to make labels for the job, we will be able to make a shipping label for the job (these buttons for shipping can be placeholders) but if you can connect the Customer -> quote -> Job details flow so they have a clean journey