Generated by Spark: Can we add filtering to the quote/job | filter by date / status
Can we add Address to the customer details page?

Can we re-engineer the "Jobs board page" the inline details are PERFECT, but the kanbaan board style doesnt work for this, a list would be better

Can we setup a search feature at the top: searches by customer name / Company name / job nickname

Can we create a "settings" on the left nav, this is where we can control things like the overall theme and other backend features

Can we create a "Reports" page which is linked to Customer data + sales data?

the "export csv" doesnt need to be on EACH page, we can move that feature into our settings, if we ever need to "export" we can easily export, customers, orders etc