Generated by Spark: Quotes page:
Can we change the "Q-2025-0045" include the  "Job nickname"
Under the Job nickname can we have the "Customer Name"

Can we create a better way to change the status of quotes from the "Quotes list page"

______________

Jobs Page:

Can we create a better way to change the status of quotes from the "job list page"

Can we add production scheduling calendar to visualize job timelines and capacity.

_______________

Can we implement automated customer communication and quote reminders. (We use outlook for email, resend, n8n)

__________________

Can we add a feature to our Line items where if we upload a file we get a visual on a mockup & it includes the imprint size (reference attatched)

___________________

In our job detail can we bring the status bar lower and not take as much space
Can we get the mockup as the first thing and then next other it the customer + date, under that will be the line item details & status